Configure mod in another file
