--- src/main/java/com/peakyhell/skytrack/commands/CommandsConfig.java
@@ -1,10 +1,0 @@
-package com.peakyhell.skytrack.commands;
-
-public class CommandsConfig {
-
-    public static void init() {
-        Hello.init();
-        ConfigScreen.init();
-    }
-
-}
